Move most parsing functions to Parser.java
