+++ src/main/java/Deadline.java
@@ -14,6 +14,7 @@ public class Deadline extends Task {
 
     @Override
     public String toString() {
-        return "[D]" + super.toString() + String.format(" (by: %s)", endTime.format(DateTimeFormatter.ofPattern("MMM d yyyy")));
+        return "[D]" + super.toString() + String.format(" (by: %s)",
+                                                        endTime.format(DateTimeFormatter.ofPattern(Parser.DATE_OUTPUT_PATTERN)));
     }
 }
