--- src/main/java/parser.java
@@ -1,3 +1,0 @@
-public class parser {
-
-}
